src/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-registration-redux==2.9
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.7.7
pytz==2021.1
sqlparse==0.4.1
